+++ util-core/src/main/java/com/twitter/io/Writers.java
@@ -7,7 +7,9 @@ import java.io.OutputStream;
  * Java APIs for Writer.
  *
  * @see com.twitter.io.Writer
+ * @deprecated This will no longer be necessary when 2.11 support is dropped. 2019-10-04
  */
+@Deprecated
 public final class Writers {
 
   private Writers() { throw new IllegalStateException(); }